Add files via upload
This is a Car Price Prediction Using CarDekho App's Dataset. I have created a new environment on Anaconda's command prompt and then executed the codes on Jupyter Notebook. This project helps in understanding the prices of the used cars in India, how old are they and what should be the price currently in the year 2021.

Python version: 3.7
Libraries used: pandas, numpy, matplotlib.pyplot, sklearn (for Randomized SearchCV), seaborn

I have also created a pickle file in the end for later purpose for deployment of the project.